feat 69 Data and domain done, struggling on presentation

+++ app/src/main/java/edu/iesam/meceiot/features/pantallasensor/domain/Sensor.kt
@@ -4,8 +4,8 @@ data class Sensor(
     val id: Int,
     val nombre: String,
     val nombrePanel: String,
-    val valoresX: List<Float>,
-    val valoresY: List<Float>,
+    val valoresX: List<Long>,
+    val valoresY: List<Int>,
     val leyendaX: String,
     val leyendaY: String
 )
